requirements.txt
numpy>=1.23,<2
scipy>=1.13,<1.15
pandas>=2.2.2,<2.3
scikit-learn>=1.5,<1.6
mlflow==2.14.1
fastapi>=0.111,<0.115
uvicorn>=0.30,<0.32
pydantic>=2.7,<2.9
pyyaml>=6.0,<7
requests>=2.31,<3
joblib>=1.3,<2
prometheus-client>=0.19,<0.21
evidently==0.4.29
shap>=0.46,<0.47
matplotlib>=3.9,<3.10
plotly>=5.22,<5.23
pytest>=8.2,<8.3
ruff==0.5.6
pre-commit>=3.7,<3.8
fastparquet>=2024.2.0,<2025.1
